Moved portfolio and contact into /pages
