--- src/markups.js
@@ -1,78 +1,0 @@
-export const navBarInnerHTML = `
-<p class="nav-title pointer">
-        <a href="../index.html" style="color: inherit">Arch</a>
-      </p>
-      <ul class="menu">
-        <li class="menu-item">
-          <a
-            class="anchor-padding-top-bottom"
-            href="pages/about-us.html"
-            style="color: inherit"
-            >Portfolio</a
-          >
-        </li>
-        <li class="menu-item">
-          <a
-            class="anchor-padding-top-bottom"
-            href="pages/about-us.html"
-            style="color: inherit"
-            >About Us</a
-          >
-        </li>
-        <li class="menu-item">
-          <a
-            class="anchor-padding-top-bottom"
-            href="index.html"
-            style="color: inherit"
-            >Contact</a
-          >
-        </li>
-      </ul>
-      <div class="dark-overlay"></div>
-      <div class="hamburger">
-        <i class="fas fa-bars"></i>
-      </div>
-`;
-
-export const footerInnerHTML = `
-<div class="footer-logo">
-        <p>arch</p>
-      </div>
-      <div class="footer-main-container">
-        <ul>
-          <li>
-            <a
-              class="anchor-padding-top-bottom"
-              href="index.html"
-              style="color: inherit"
-              >Portfolio</a
-            >
-          </li>
-          <li>
-            <a
-              class="anchor-padding-top-bottom"
-              href="pages/about-us.html"
-              style="color: inherit"
-              >About Us</a
-            >
-          </li>
-          <li>
-            <a
-              class="anchor-padding-top-bottom"
-              href="index.html"
-              style="color: inherit"
-              >Contact</a
-            >
-          </li>
-        </ul>
-        <div class="footer-button-container">
-          <a href="index.html">
-            <button class="button">
-              <p>See Our Portfolio</p>
-              <i class="fas fa-arrow-right"></i>
-            </button>
-          </a>
-        </div>
-      </div>  
-
-`;
