+++ src/pages/contact/index.html
@@ -5,5 +5,5 @@
 
 <p>contact</p>
 
-<script type="module" src="../main.js"></script>
+<script type="module" src="../../main.js"></script>
 <script type="module" src="./contact.js"></script>
