+++ src/pages/portfolio/constants.js
@@ -1,0 +1,136 @@
+import seraphMobile from "../../assets/portfolio/tablet/image-seraph.jpg";
+import seraphTablet from "../../assets/portfolio/mobile/image-seraph.jpg";
+import seraphDesktop from "../../assets/portfolio/desktop/image-seraph.jpg";
+
+import eeboxMobile from "../../assets/portfolio/mobile/image-eebox.jpg";
+import eeboxTablet from "../../assets/portfolio/tablet/image-eebox.jpg";
+import eeboxDesktop from "../../assets/portfolio/desktop/image-eebox.jpg";
+
+import federalMobile from "../../assets/portfolio/mobile/image-federal.jpg";
+import federalTablet from "../../assets/portfolio/tablet/image-federal.jpg";
+import federalDesktop from "../../assets/portfolio/desktop/image-federal.jpg";
+
+import desolMobile from "../../assets/portfolio/mobile/image-del-sol.jpg";
+import desolTablet from "../../assets/portfolio/tablet/image-del-sol.jpg";
+import desolDesktop from "../../assets/portfolio/desktop/image-del-sol.jpg";
+
+import prototypeMobile from "../../assets/portfolio/mobile/image-prototype.jpg";
+import prototypeTablet from "../../assets/portfolio/tablet/image-prototype.jpg";
+import prototypeDesktop from "../../assets/portfolio/desktop/image-prototype.jpg";
+
+import b228Mobile from "../../assets/portfolio/mobile/image-228b.jpg";
+import b228Tablet from "../../assets/portfolio/tablet/image-228b.jpg";
+import b228Desktop from "../../assets/portfolio/desktop/image-228b.jpg";
+
+import edelweissMobile from "../../assets/portfolio/mobile/image-edelweiss.jpg";
+import edelweissTablet from "../../assets/portfolio/tablet/image-edelweiss.jpg";
+import edelweissDesktop from "../../assets/portfolio/desktop/image-edelweiss.jpg";
+
+import netcryMobile from "../../assets/portfolio/mobile/image-netcry.jpg";
+import netcryTablet from "../../assets/portfolio/tablet/image-netcry.jpg";
+import netcryDesktop from "../../assets/portfolio/desktop/image-netcry.jpg";
+
+import hypersMobile from "../../assets/portfolio/mobile/image-hypers.jpg";
+import hypersTablet from "../../assets/portfolio/tablet/image-hypers.jpg";
+import hypersDesktop from "../../assets/portfolio/desktop/image-hypers.jpg";
+
+import sxivMobile from "../../assets/portfolio/mobile/image-sxiv.jpg";
+import sxivTablet from "../../assets/portfolio/tablet/image-sxiv.jpg";
+import sxivDesktop from "../../assets/portfolio/desktop/image-sxiv.jpg";
+
+import trinityMobile from "../../assets/portfolio/mobile/image-trinity.jpg";
+import trinityTablet from "../../assets/portfolio/tablet/image-trinity.jpg";
+import trinityDesktop from "../../assets/portfolio/desktop/image-trinity.jpg";
+
+import paramourMobile from "../../assets/portfolio/mobile/image-paramour.jpg";
+import paramourTablet from "../../assets/portfolio/tablet/image-paramour.jpg";
+import paramourDesktop from "../../assets/portfolio/desktop/image-paramour.jpg";
+
+const cardsConfig = [
+  {
+    name: "Seraph Station",
+    subtitle: "September 2019",
+    bgImageUrlTablet: seraphMobile,
+    bgImageUrlMobile: seraphTablet,
+    bgImageUrlDesktop: seraphDesktop,
+  },
+  {
+    name: "Eebox Building",
+    subtitle: "August 2017",
+    bgImageUrlMobile: eeboxMobile,
+    bgImageUrlTablet: eeboxTablet,
+    bgImageUrlDesktop: eeboxDesktop,
+  },
+  {
+    name: "Federal II Tower",
+    subtitle: "March 2017",
+    bgImageUrlMobile: federalMobile,
+    bgImageUrlTablet: federalTablet,
+    bgImageUrlDesktop: federalDesktop,
+  },
+  {
+    name: "Project Del Sol",
+    subtitle: "January 2016",
+    bgImageUrlMobile: desolMobile,
+    bgImageUrlTablet: desolTablet,
+    bgImageUrlDesktop: desolDesktop,
+  },
+  {
+    name: "Le Prototype",
+    subtitle: "October 2015",
+    bgImageUrlMobile: prototypeMobile,
+    bgImageUrlTablet: prototypeTablet,
+    bgImageUrlDesktop: prototypeDesktop,
+  },
+  {
+    name: "228B Tower",
+    subtitle: "April 2015",
+    bgImageUrlMobile: b228Mobile,
+    bgImageUrlTablet: b228Tablet,
+    bgImageUrlDesktop: b228Desktop,
+  },
+  {
+    name: "Grand Edelweiss Hotel",
+    subtitle: "September 2019",
+    bgImageUrlMobile: edelweissMobile,
+    bgImageUrlTablet: edelweissTablet,
+    bgImageUrlDesktop: edelweissDesktop,
+  },
+  {
+    name: "Netcry Tower",
+    subtitle: "August 2012",
+    bgImageUrlMobile: netcryMobile,
+    bgImageUrlTablet: netcryTablet,
+    bgImageUrlDesktop: netcryDesktop,
+  },
+  {
+    name: "Hypers",
+    subtitle: "January 2012",
+    bgImageUrlMobile: hypersMobile,
+    bgImageUrlTablet: hypersTablet,
+    bgImageUrlDesktop: hypersDesktop,
+  },
+  {
+    name: "SXIV Tower",
+    subtitle: "March 2011",
+    bgImageUrlMobile: sxivMobile,
+    bgImageUrlTablet: sxivTablet,
+    bgImageUrlDesktop: sxivDesktop,
+  },
+  {
+    name: "Trinity Bank Tower",
+    subtitle: "September 2010",
+    bgImageUrlMobile: trinityMobile,
+    bgImageUrlTablet: trinityTablet,
+    bgImageUrlDesktop: trinityDesktop,
+  },
+  {
+    name: "Project Paramour",
+    subtitle: "February 2008",
+    bgImageUrlMobile: paramourMobile,
+    bgImageUrlTablet: paramourTablet,
+    bgImageUrlDesktop: paramourDesktop,
+  },
+];
+
+export default cardsConfig;
