+++ src/pages/portfolio/index.html
@@ -16,5 +16,5 @@
       -->
 </section>
 
-<script type="module" src="../main.js"></script>
+<script type="module" src="../../main.js"></script>
 <script type="module" src="./portfolio.js"></script>
