+++ src/pages/portfolio/portfolio.scss
@@ -1,7 +1,7 @@
-@import "../styles/colors";
-@import "../styles/typography";
-@import "../styles/variables";
-@import "../styles/mixins";
+@import "../../styles/colors";
+@import "../../styles/typography";
+@import "../../styles/variables";
+@import "../../styles/mixins";
 
 .portfolio {
   margin-top: $navbar-height * 2;
